Data Collection and Processing Module/requirements.txt
Flask==2.3.0
requests==2.31.0
pandas==2.1.1
numpy==1.24.4
scikit-learn==1.3.0
sqlalchemy==2.0.19
psycopg2-binary==2.9.8
apache-kafka==2.1.0
pyyaml==6.1
geopandas==0.13.0
